Move Floating and Util to Internal Package
