+++ docking-api/src/io/github/andrewauclair/moderndocking/internal/DockableToolbar.java
@@ -26,10 +26,10 @@ import io.github.andrewauclair.moderndocking.api.DockingAPI;
 import io.github.andrewauclair.moderndocking.api.RootDockingPanelAPI;
 import io.github.andrewauclair.moderndocking.ui.DockingSettings;
 import io.github.andrewauclair.moderndocking.ui.ToolbarLocation;
-import io.github.andrewauclair.moderndocking.util.CombinedIcon;
-import io.github.andrewauclair.moderndocking.util.RotatedIcon;
-import io.github.andrewauclair.moderndocking.util.TextIcon;
-import io.github.andrewauclair.moderndocking.util.UnselectableButtonGroup;
+import io.github.andrewauclair.moderndocking.internal.util.CombinedIcon;
+import io.github.andrewauclair.moderndocking.internal.util.RotatedIcon;
+import io.github.andrewauclair.moderndocking.internal.util.TextIcon;
+import io.github.andrewauclair.moderndocking.internal.util.UnselectableButtonGroup;
 import java.awt.Color;
 import java.awt.GridBagConstraints;
 import java.awt.GridBagLayout;
