+++ docking-api/src/io/github/andrewauclair/moderndocking/internal/DockableWrapper.java
@@ -24,8 +24,8 @@ package io.github.andrewauclair.moderndocking.internal;
 import io.github.andrewauclair.moderndocking.Dockable;
 import io.github.andrewauclair.moderndocking.Property;
 import io.github.andrewauclair.moderndocking.api.DockingAPI;
-import io.github.andrewauclair.moderndocking.floating.DisplayPanelFloatListener;
-import io.github.andrewauclair.moderndocking.floating.FloatListener;
+import io.github.andrewauclair.moderndocking.internal.floating.DisplayPanelFloatListener;
+import io.github.andrewauclair.moderndocking.internal.floating.FloatListener;
 import io.github.andrewauclair.moderndocking.ui.DockingHeaderUI;
 import io.github.andrewauclair.moderndocking.ui.HeaderController;
 import io.github.andrewauclair.moderndocking.ui.HeaderModel;
