+++ docking-api/src/io/github/andrewauclair/moderndocking/internal/DockedTabbedPanel.java
@@ -25,9 +25,9 @@ import io.github.andrewauclair.moderndocking.Dockable;
 import io.github.andrewauclair.moderndocking.DockableTabPreference;
 import io.github.andrewauclair.moderndocking.DockingRegion;
 import io.github.andrewauclair.moderndocking.api.DockingAPI;
-import io.github.andrewauclair.moderndocking.floating.DockedTabbedPanelFloatListener;
-import io.github.andrewauclair.moderndocking.floating.FloatListener;
-import io.github.andrewauclair.moderndocking.floating.Floating;
+import io.github.andrewauclair.moderndocking.internal.floating.DockedTabbedPanelFloatListener;
+import io.github.andrewauclair.moderndocking.internal.floating.FloatListener;
+import io.github.andrewauclair.moderndocking.internal.floating.Floating;
 import io.github.andrewauclair.moderndocking.settings.Settings;
 import io.github.andrewauclair.moderndocking.ui.DockingSettings;
 import java.awt.BorderLayout;
