+++ docking-api/src/io/github/andrewauclair/moderndocking/internal/DockingInternal.java
@@ -29,7 +29,7 @@ import io.github.andrewauclair.moderndocking.api.RootDockingPanelAPI;
 import io.github.andrewauclair.moderndocking.exception.DockableNotFoundException;
 import io.github.andrewauclair.moderndocking.exception.DockableRegistrationFailureException;
 import io.github.andrewauclair.moderndocking.exception.RootDockingPanelRegistrationFailureException;
-import io.github.andrewauclair.moderndocking.floating.Floating;
+import io.github.andrewauclair.moderndocking.internal.floating.Floating;
 import io.github.andrewauclair.moderndocking.ui.DefaultHeaderUI;
 import io.github.andrewauclair.moderndocking.ui.DockingHeaderUI;
 import io.github.andrewauclair.moderndocking.ui.HeaderController;
@@ -39,14 +39,12 @@ import java.awt.Window;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.util.Arrays;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.function.BiFunction;
 import java.util.stream.Collectors;
-import javax.swing.JComponent;
 import javax.swing.JDialog;
 import javax.swing.JFrame;
 import javax.swing.SwingUtilities;
