+++ docking-api/src/io/github/andrewauclair/moderndocking/internal/util/RotatedIcon.java
@@ -1,4 +1,4 @@
-package io.github.andrewauclair.moderndocking.util;
+package io.github.andrewauclair.moderndocking.internal.util;
 
 import java.awt.Component;
 import java.awt.Graphics;
